#87 - done comparision with int values

+++ src/com/lftechnology/java/training/alina/collections/generics/wildcards/Customer.java
@@ -38,7 +38,7 @@ public class Customer {
         this.address = address;
     }
 
-    public float getBill() {
+    public int getBill() {
 
         return bill;
     }
